Patcher: use wine_get_host_version to detect Linux and macOS from Wine.
The previous mechanism to detect the host operating system when running
via Wine relied on checking the existence of specific files in the host
system. Some Wine distributions, like CrossOver, use sandboxing and
filesystem isolation for improved security, making the check fail.
The new approach is to call wine_get_host_version, exposed by Wine in
ntdll.dll.
