+++ ladxhd_patcher_source_code/LADXHD_Patcher.csproj
@@ -81,7 +81,7 @@
     <Compile Include="Program\Config.cs" />
     <Compile Include="Program\Forms.cs" />
     <Compile Include="Program\Functions.cs" />
-    <Compile Include="Program\HostEnvironment.cs" />
+    <Compile Include="Program\WineHost.cs" />
     <Compile Include="Program\XDelta3.cs" />
     <Compile Include="Program\Utilities.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
